pyproject.toml
[project]
name = "square"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "python-dateutil",
    "rapidfuzz>=3.14.3",
    "squareup>=43.1.2.20250924",
]
